fix(playground): only connect browser-stream WebSocket for agents with browser tools
BrowserSessionProvider auto-connected to /browser/:agentId/stream for every
agent chat session, regardless of whether the agent had a browser toolset.
For agents without browser tools, the WebSocket would open against a route
with no toolset to bind to and trigger up to 5 exponential-backoff reconnect
attempts. Each status flip ('connecting' -> 'disconnected' -> 'connecting')
invalidated the memoized context value, rerendering every useBrowserSession
consumer and thrashing surrounding UI state (sidebars/panels collapsing).

Add an 'enabled' prop on BrowserSessionProvider that gates connect(), the
auto-connect effect, and the visibility-change reconnect listener. Wire it
from all three mount sites (agent chat, session view, builder test chat)
based on agent.browserTools?.length — the agent details endpoint already
exposes this field.

Split BrowserSessionProvider into its own file so browser-session-context.tsx
exports only the context object, types, and useBrowserSession hook — no
components — which unblocks the react-refresh/only-export-components lint
rule that lint-staged enforces with --max-warnings=0.
