+++ packages/playground/src/domains/agents/context/index.tsx
@@ -1,6 +1,8 @@
+/* eslint-disable react-refresh/only-export-components */
 export * from './agent-context';
 export * from './agent-working-memory-context';
 export * from './browser-session-context';
+export * from './browser-session-provider';
 export * from './browser-tool-calls-context';
 export * from './agent-observational-memory-context';
 export * from './activated-skills-context';
